Minor changed to the threads. Initialize fork solution
